# web/docs/swagger.yaml
basePath: /
definitions:
  errorinterface.ErrorResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
  keycloak.PolicyRepresentation:
    properties:
      decisionStrategy:
        type: string
      description:
        type: string
      id:
        type: string
      logic:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  keycloak.ResourceRepresentation:
    properties:
      _id:
        type: string
      attributes:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      displayName:
        type: string
      icon_uri:
        type: string
      name:
        type: string
      owner:
        additionalProperties:
          type: string
        type: object
      scopes:
        items:
          $ref: '#/definitions/keycloak.ScopeRepresentation'
        type: array
      type:
        type: string
      uris:
        items:
          type: string
        type: array
    type: object
  keycloak.ScopeRepresentation:
    properties:
      displayName:
        type: string
      iconUri:
        type: string
      id:
        type: string
      name:
        type: string
      policies:
        items:
          $ref: '#/definitions/keycloak.PolicyRepresentation'
        type: array
      resources:
        items:
          $ref: '#/definitions/keycloak.ResourceRepresentation'
        type: array
    type: object
...
  oauth.ResourceRequest:
    properties:
      _id:
        type: string
      attributes:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      displayName:
        type: string
      icon_uri:
        type: string
      name:
        type: string
      owner:
        additionalProperties:
          type: string
        type: object
      scopes:
        items:
          $ref: '#/definitions/keycloak.ScopeRepresentation'
        type: array
      type:
        type: string
      uris:
        items:
          type: string
        type: array
    type: object
  oauth.ResourceResponse:
    properties:
      _id:
        type: string
      attributes:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      displayName:
        type: string
      icon_uri:
        type: string
      name:
        type: string
      owner:
        additionalProperties:
          type: string
        type: object
      scopes:
        items:
          $ref: '#/definitions/keycloak.ScopeRepresentation'
        type: array
      type:
        type: string
      uris:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8002
info:
  contact: {}
paths:
  /{workspace_id}/{project_id}/oauth/{resource_id}:
    put:
      consumes:
      - application/json
      description: Updates an existing Keycloak resource
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      - description: Resource Details
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/oauth.ResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/oauth.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Update a resource
      tags:
      - resources
  /{workspace_id}/{project_id}/oauth/resources:
    get:
      consumes:
      - application/json
      description: Lists all Keycloak resources with optional filtering
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      - description: Filter by URI
        in: query
        name: uri
        type: string
      - description: Filter by owner
        in: query
        name: owner
        type: string
      - description: Filter by scope
        in: query
        name: scope
        type: string
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/oauth.ResourceResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: List resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: Creates a new Keycloak resource
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Resource Details
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/oauth.ResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/oauth.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Create a new resource
      tags:
      - resources
  /{workspace_id}/{project_id}/oauth/resources/{resource_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a Keycloak resource
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Delete a resource
      tags:
      - resources
    get:
      consumes:
      - application/json
      description: Retrieves a Keycloak resource by ID
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/oauth.ResourceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Get a resource
      tags:
      - resources
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"